requirements.txt
accelerate==0.16.0
alembic==1.9.2
argon2_cffi==21.3.0
broadcaster[redis]==0.2.0
diffusers==0.12.1
dramatiq[redis]==1.14.0
email-validator==1.3.1
fastapi==0.89.1
httpx==0.23.3
minio==7.1.13
motor==3.1.1
numpy==1.24.2
pandas==1.5.3
passlib==1.7.4
Pillow==9.4.0
python-multipart==0.0.5
scikit-learn==1.2.1
sqlalchemy[asyncio,mypy]==2.0.1
starlette-csrf==1.4.5
transformers[torch]==4.26.0

uvicorn[standard]

aiosqlite
asgi_lifespan
black
httpx-ws
mypy
pandas-stubs
pytest
pytest-asyncio
pytest-cov
ruff
types-passlib
types-Pillow
